move enterprise/internal/compute to internal/compute
