+++ enterprise/internal/insights/query/querybuilder/BUILD.bazel
@@ -11,8 +11,8 @@ go_library(
     importpath = "github.com/sourcegraph/sourcegraph/enterprise/internal/insights/query/querybuilder",
     visibility = ["//enterprise:__subpackages__"],
     deps = [
-        "//enterprise/internal/compute",
         "//enterprise/internal/insights/types",
+        "//internal/compute",
         "//internal/search/client",
         "//internal/search/query",
         "//internal/search/repos",
