+++ internal/compute/BUILD.bazel
@@ -14,7 +14,7 @@ go_library(
         "template.go",
         "text_result.go",
     ],
-    importpath = "github.com/sourcegraph/sourcegraph/enterprise/internal/compute",
+    importpath = "github.com/sourcegraph/sourcegraph/internal/compute",
     visibility = ["//enterprise:__subpackages__"],
     deps = [
         "//internal/authz",
