+++ internal/compute/client/BUILD.bazel
@@ -7,10 +7,10 @@ go_library(
         "compute_text_client.go",
         "match_context_client.go",
     ],
-    importpath = "github.com/sourcegraph/sourcegraph/enterprise/internal/compute/client",
+    importpath = "github.com/sourcegraph/sourcegraph/internal/compute/client",
     visibility = ["//enterprise:__subpackages__"],
     deps = [
-        "//enterprise/internal/compute",
+        "//internal/compute",
         "//internal/search/streaming/api",
         "//internal/search/streaming/http",
         "//lib/errors",
@@ -26,7 +26,7 @@ go_test(
     ],
     embed = [":client"],
     deps = [
-        "//enterprise/internal/compute",
+        "//internal/compute",
         "//internal/search/streaming/api",
         "//internal/search/streaming/http",
         "@com_github_hexops_autogold_v2//:autogold",
