+++ internal/compute/client/match_context_client_test.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/hexops/autogold/v2"
 
-	"github.com/sourcegraph/sourcegraph/enterprise/internal/compute"
+	"github.com/sourcegraph/sourcegraph/internal/compute"
 	"github.com/sourcegraph/sourcegraph/internal/search/streaming/api"
 	"github.com/sourcegraph/sourcegraph/internal/search/streaming/http"
 )
