+++ internal/compute/query_test.go
