+++ internal/compute/result.go
