+++ internal/compute/testdata/Test_matchOnly/match_only#02.golden
