+++ internal/compute/testdata/Test_matchOnly/match_only#03.golden
