+++ internal/compute/testdata/Test_matchOnly/match_only#04.golden
