+++ internal/compute/testdata/Test_matchOnly/match_only#05.golden
