+++ internal/compute/testdata/Test_matchOnly/match_only#06.golden
