+++ internal/compute/text_result.go
